Add DMA-BUF support
Support cases where the rendering GPU isn't the same as the displaying
GPU. This is done by allowing the drm and gbm structure to be based on
top of different file descriptors. If that's the case, buffers will be
exported from gbm to a DMA-BUF file descriptor and reimported into the
DRM device for display.

To use this, specify both devices on the command-line, like so:

	$ kmscube /dev/dri/card1 /dev/dri/card0

The first parameter specifies the path to the display device, whereas
the second parameter points to the rendering device.

As a bonus, specifying only one argument allows a specific device to be
used for display and rendering in setups where two or more devices are
available. Furthermore the DMA-BUF functionality can also be tested on
setups with only a single GPU by specifying the same path for both
arguments.

Signed-off-by: Thierry Reding <treding@nvidia.com>